+++ packages/data-explorer/src/components/dataView/CustomCardContent.vue
@@ -1,14 +1,34 @@
 <template>
-  <div>Custom card content</div>
+  <div class="custom-card">
+    <slot name="shopping-button"></slot>
+    <v-runtime-template :template="template"></v-runtime-template>
+  </div>
 </template>
 
 <script>
-export default {
-  name: 'CustomCardContent'
+import VRuntimeTemplate from 'v-runtime-template'
 
+export default {
+  name: 'CustomCardContent',
+  components: { VRuntimeTemplate },
+  props: {
+    id: {
+      type: String,
+      required: true
+    },
+    record: {
+      type: Object,
+      required: true
+    },
+    customCode: {
+      type: String,
+      required: true
+    }
+  },
+  computed: {
+    template: function () {
+      return '<div>' + this.customCode.trim() + '</div>'
+    }
+  }
 }
 </script>
-
-<style scoped>
-
-</style>