+++ packages/data-explorer/src/store/state.ts
@@ -2,20 +2,27 @@ import ApplicationState from '@/types/ApplicationState'
 
 const state: ApplicationState = {
   toast: null,
-  tableName: 'it_emx_datatypes_TypeTest',
+  tableName: 'root_hospital_patients',
   tableData: null,
   tableMeta: null,
   dataDisplayLayout: 'CardView',
   defaultEntityData: null,
   entityMetaRefs: {},
-  hideFilters: false,
   showShoppingCart: false,
   shoppedEntityItems: [],
   tableSettings: {
     settingsTable: 'de_dataexplorer_table_settings',
     settingsRowId: null,
     collapseLimit: 5,
+    customCardCode: null,
+    customCardAttrs: '',
     isShop: false
+  },
+  filters: {
+    hideSidebar: false,
+    definition: [],
+    shown: [],
+    selections: {}
   }
 }