+++ packages/data-explorer/src/views/ClipboardView.vue
@@ -34,7 +34,7 @@
 <script>
 import TableRow from '../components/dataView/TableRow'
 import TableHeader from '../components/dataView/TableHeader'
-import { mapState } from 'vuex'
+import { mapState, mapMutations, mapActions } from 'vuex'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faShoppingBag, faChevronLeft } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
@@ -44,7 +44,7 @@ export default {
   name: 'ClipboardView',
   components: { TableRow, TableHeader, FontAwesomeIcon },
   computed: {
-    ...mapState(['tableMeta', 'shoppedEntityItems', 'tableData']),
+    ...mapState(['tableMeta', 'shoppedEntityItems', 'tableData', 'tableName']),
     idAttribute () {
       return this.tableMeta.idAttribute
     },
@@ -56,6 +56,8 @@ export default {
     }
   },
   methods: {
+    ...mapActions(['fetchTableViewData']),
+    ...mapMutations(['setShowShoppingCart', 'setHideFilters']),
     getEntityId (entity) {
       return entity[this.idAttribute].toString()
     },
@@ -63,9 +65,12 @@ export default {
       return this.shoppedEntityItems.includes(this.getEntityId(entity))
     },
     closeShoppingCart () {
-      this.$store.commit('setShowShoppingCart', false)
-      this.$store.commit('setHideFilters', false)
+      this.setShowShoppingCart(false)
+      this.setHideFilters(false)
     }
+  },
+  mounted: function () {
+    this.fetchTableViewData({ tableName: this.tableName })
   }
 }
 </script>