+++ packages/data-explorer/src/views/SelectLayoutView.vue
@@ -1,11 +1,11 @@
 <template>
-  <div class="mt-3 entity-table" v-if="tableData && tableData.items.length > 0 && tableMeta">
-    <div v-if="tableData.items.length === 0" class="alert alert-warning">
+  <div class="mt-3 entity-table" v-show="tableData && tableData.items && tableData.items.length > 0 && tableMeta">
+    <div v-if="tableData && tableData.items && tableData.items.length === 0" class="alert alert-warning">
       {{ 'dataexplorer_empty_table' | i18n}}
     </div>
     <component
       :is="dataDisplayLayout"
-      :entitiesToShow="tableData.items"
+      :entitiesToShow="(tableData && tableData.items) ? tableData.items : []"
     />
   </div>
 </template>