+++ packages/data-explorer/tests/unit/mocks/tableDataResponseMock.ts
@@ -1,0 +1,12 @@
+export default {
+  links: {
+    self: 'https://mocks-for-jocks.com/data/it_emx_datatypes_TypeTest/1'
+  },
+  items: [{
+    links: '',
+    data: {
+      id: 1,
+      label: 'my label table data'
+    }
+  }]
+}