+++ packages/metadata-manager/src/components/MetadataManagerEntityEditForm.vue
@@ -47,8 +47,8 @@
             <input v-model="label" class="form-control" type="text"
                    :placeholder="$t('entity-edit-form-label-placeholder')">
             <div class="input-group-append" style="z-index: -1"><!-- -1, because otherwise append overlaps entity type dropdown -->
-              <button @click="showLabelLanguageInputs = !showLabelLanguageInputs" class="btn btn-outline-secondary"><i
-                class="fa fa-language fa-lg"></i></button>
+              <button @click="showLabelLanguageInputs = !showLabelLanguageInputs" class="btn btn-outline-secondary" aria-label="translate labels"><i
+                class="fa fa-language fa-lg" aria-hidden="true"></i></button>
             </div>
           </div>
         </div>
@@ -72,8 +72,8 @@
             <input v-model="description" class="form-control" type="text"
                    :placeholder="$t('entity-edit-form-description-placeholder')">
             <div class="input-group-append" style="z-index: -1"><!-- -1, because otherwise append overlaps entity type dropdown -->
-              <button @click="showDescriptionLanguageInputs = !showDescriptionLanguageInputs" class="btn btn-outline-secondary"><i
-                class="fa fa-language fa-lg"></i></button>
+              <button @click="showDescriptionLanguageInputs = !showDescriptionLanguageInputs" class="btn btn-outline-secondary" aria-label="translate descriptions"><i
+                class="fa fa-language fa-lg" aria-hidden="true"></i></button>
             </div>
           </div>
         </div>