+++ packages/questionnaires/test/e2e/specs/questionnaire-walkthrough.spec.js
@@ -10,6 +10,8 @@ const previousChapterButton = '#prev-chapter-btn'
 const currentChapterSpan = '#current-chapter-label'
 const submitButton = '#submit-questionnaire-btn'
 
+const timeAllowedForProgressBarUpdate = 2000
+
 module.exports = {
   before: function (browser) {
     browser.url(browser.globals.devServerURL)
@@ -34,12 +36,12 @@ module.exports = {
     // ======= tests =======
     browser.expect.element(tableSelector).to.be.present
 
-    browser.expect.element(firstRowSelector + ' > ' + firstColumnSelector).text.to.equal('Questionnaire #1')
-    browser.expect.element(secondRowSelector + ' > ' + firstColumnSelector).text.to.equal('Questionnaire #2')
-    browser.expect.element(thirdRowSelector + ' > ' + firstColumnSelector).text.to.equal('Questionnaire #3')
-    browser.expect.element(firstRowSelector + ' > ' + secondColumnSelector).text.to.equal('Not started yet')
-    browser.expect.element(secondRowSelector + ' > ' + secondColumnSelector).text.to.equal('Open')
-    browser.expect.element(thirdRowSelector + ' > ' + secondColumnSelector).text.to.equal('Submitted')
+    browser.expect.element(firstRowSelector + ' > ' + firstColumnSelector).text.to.contains('Questionnaire #1')
+    browser.expect.element(secondRowSelector + ' > ' + firstColumnSelector).text.to.contains('Questionnaire #2')
+    browser.expect.element(thirdRowSelector + ' > ' + firstColumnSelector).text.to.contains('Questionnaire #3')
+    browser.expect.element(firstRowSelector + ' > ' + secondColumnSelector).text.to.contains('Not started yet')
+    browser.expect.element(secondRowSelector + ' > ' + secondColumnSelector).text.to.contains('Open')
+    browser.expect.element(thirdRowSelector + ' > ' + secondColumnSelector).text.to.contains('Submitted')
   },
 
   'should load start page when button is clicked': function (browser) {
@@ -61,11 +63,10 @@ module.exports = {
     browser.expect.element(questionnaireLabelElement).to.be.present
     browser.expect.element(questionnaireDescriptionElement).to.be.present
 
-    browser.expect.element(backToQuestionnairesListButton).text.to.equal('Back to questionnaires list')
-    browser.expect.element(startQuestionnaireButton).text.to.equal('Start questionnaire')
-    browser.expect.element(questionnaireLabelElement).text.to.equal('Questionnaire #1')
-    browser.expect.element(questionnaireDescriptionElement).text.to.equal('This is the first mocked questionnaire response\n' +
-      'This is a not started questionnaire used for showing the chapters')
+    browser.expect.element(backToQuestionnairesListButton).text.to.contains('Back to questionnaires list')
+    browser.expect.element(startQuestionnaireButton).text.to.contains('Start questionnaire')
+    browser.expect.element(questionnaireLabelElement).text.to.contains('Questionnaire #1')
+    browser.expect.element(questionnaireDescriptionElement).text.to.contains('This is the first mocked questionnaire response')
   },
 
   'should load the first chapter when the start questionnaire button is clicked': function (browser) {
@@ -88,12 +89,12 @@ module.exports = {
     browser.expect.element(firstChapterListItem).to.be.present
     browser.expect.element(secondChapterListItem).to.be.present
 
-    browser.expect.element(backToStartButton).text.to.equal('Back to start')
-    browser.expect.element(nextChapterButton).text.to.equal('Next chapter')
-    browser.expect.element(currentChapterSpan).text.to.equal('Chapter 1 of 2')
-    browser.expect.element(chapterNavigationHeader).text.to.equal('Chapters')
-    browser.expect.element(firstChapterListItem).text.to.equal('Chapter #1 - Personal information')
-    browser.expect.element(secondChapterListItem).text.to.equal('Chapter #2 - Professional questions')
+    browser.expect.element(backToStartButton).text.to.contains('Back to start')
+    browser.expect.element(nextChapterButton).text.to.contains('Next chapter')
+    browser.expect.element(currentChapterSpan).text.to.contains('Chapter 1 of 2')
+    browser.expect.element(chapterNavigationHeader).text.to.contains('Chapters')
+    browser.expect.element(firstChapterListItem).text.to.contains('Chapter #1 - Personal information')
+    browser.expect.element(secondChapterListItem).text.to.contains('Chapter #2 - Professional questions')
   },
 
   'should update progress in chapter navigation list when filling in questions': function (browser) {
@@ -111,16 +112,20 @@ module.exports = {
     browser.expect.element(chapterOneProgressBar).to.have.attribute('aria-valuenow').which.contains('0')
 
     browser.setValue(chapterOneNameQuestion, 'Nightwatch test')
+    browser.pause(timeAllowedForProgressBarUpdate)
     browser.expect.element(chapterOneProgressBar).to.have.attribute('aria-valuenow').which.contains('33')
 
     browser.setValue(chapterOneAgeQuestion, 20)
+    browser.pause(timeAllowedForProgressBarUpdate)
     browser.expect.element(chapterOneProgressBar).to.have.attribute('aria-valuenow').which.contains('67')
 
     // this question is nillable and should not update progress
     browser.click(chapterOneBoolQuestion)
+    browser.pause(timeAllowedForProgressBarUpdate)
     browser.expect.element(chapterOneProgressBar).to.have.attribute('aria-valuenow').which.contains('67')
 
     browser.setValue(chapterOneRequiredQuestion, 'A nightwatch generated text')
+    browser.pause(timeAllowedForProgressBarUpdate)
     browser.expect.element(chapterOneProgressBar).to.have.attribute('aria-valuenow').which.contains('100')
 
     browser.expect.element(chapterOneProgressBar).to.have.attribute('class').which.contains('bg-success')
@@ -140,9 +145,9 @@ module.exports = {
     browser.expect.element(currentChapterSpan).to.be.present
     browser.expect.element(submitButton).to.be.present
 
-    browser.expect.element(previousChapterButton).text.to.equal('Previous chapter')
-    browser.expect.element(currentChapterSpan).text.to.equal('Chapter 2 of 2')
-    browser.expect.element(submitButton).text.to.equal('Submit')
+    browser.expect.element(previousChapterButton).text.to.contains('Previous chapter')
+    browser.expect.element(currentChapterSpan).text.to.contains('Chapter 2 of 2')
+    browser.expect.element(submitButton).text.to.contains('Submit')
   },
 
   'should finish questionnaire and submit': function (browser) {
@@ -166,8 +171,8 @@ module.exports = {
     browser.expect.element(submissionText).to.be.present
     browser.expect.element(backToQuestionnairesListButton).to.be.present
 
-    browser.expect.element(submissionText).text.to.equal('Thank you')
-    browser.expect.element(backToQuestionnairesListButton).text.to.equal('Back to questionnaires list')
+    browser.expect.element(submissionText).text.to.contains('Thank you')
+    browser.expect.element(backToQuestionnairesListButton).text.to.contains('Back to questionnaires list')
 
     browser.click(backToQuestionnairesListButton)
     browser.expect.element(tableSelector).to.be.present