+++ packages/navigator/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/molgenis/molgenis-frontend/compare/@molgenis-ui/navigator@0.1.3...@molgenis-ui/navigator@0.1.4) (2019-10-02)
+
+
+### Bug Fixes
+
+* **navigator:** use semantic html tag ([85a9760](https://github.com/molgenis/molgenis-frontend/commit/85a9760))
+
+
+
+
+
 ## [0.1.3](https://github.com/molgenis/molgenis-frontend/compare/@molgenis-ui/navigator@0.1.2...@molgenis-ui/navigator@0.1.3) (2019-06-14)